+++ test/unit/mail_2_rmq/process_message.py
@@ -116,8 +116,8 @@ class UnitTest(unittest.TestCase):
         self.email_list2 = ["goodname@domain"]
 
     @mock.patch("mail_2_rmq.process_file", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
@@ -141,8 +141,8 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(mail_2_rmq.process_message(self.cfg, mock_log))
 
     @mock.patch("mail_2_rmq.process_from", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
@@ -166,8 +166,8 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(mail_2_rmq.process_message(self.cfg, mock_log))
 
     @mock.patch("mail_2_rmq.process_file", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
@@ -191,8 +191,8 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(mail_2_rmq.process_message(self.cfg, mock_log))
 
     @mock.patch("mail_2_rmq.process_file", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
@@ -216,8 +216,8 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(mail_2_rmq.process_message(self.cfg, mock_log))
 
     @mock.patch("mail_2_rmq.process_file", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
@@ -241,8 +241,8 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(mail_2_rmq.process_message(self.cfg, mock_log))
 
     @mock.patch("mail_2_rmq.process_file", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")
@@ -266,8 +266,8 @@ class UnitTest(unittest.TestCase):
         self.assertFalse(mail_2_rmq.process_message(self.cfg, mock_log))
 
     @mock.patch("mail_2_rmq.process_subj", mock.Mock(return_value=True))
-    @mock.patch("mail_2_rmq.get_email_addr")
-    @mock.patch("mail_2_rmq.camelize")
+    @mock.patch("mail_2_rmq.gen_libs.find_email_addr")
+    @mock.patch("mail_2_rmq.gen_libs.pascalize")
     @mock.patch("mail_2_rmq.filter_subject")
     @mock.patch("mail_2_rmq.parse_email")
     @mock.patch("mail_2_rmq.gen_class.Logger")